[TASK] Deprecate Content Object HTML
Since TYPO3 4.5 TEXT and HTML behave the same way.
HTML does not really make sense anymore and as such
could be deprecated.

After the release of 4.5 the HTML object will behave just like the TEXT object until it will be removed after the deprecation period. This is why the code has been copied from TEXT to HTML, which doesn't break the behaviour but makes it clear even codewise, that there is a redundant object.

Add a Note to the NEWS.txt.

Change-Id: I0080345efd881e273537db8797e31038d185c36c
Resolves: #28015