[FEATURE] Use caching framework in autoloader
The patch moves the initialization of the caching framework to
config_default.php, prior to the autoload registration. In effect the
autoloader can use the phpcode cache to write a full autoload registry
that can be required directly to reduce file access overhead.
As a side effect the cache system is now always enabled for all access
scopes like FE, BE, CLI and so on.

Change-Id: I7c548fbabeb6e60b8bd9c4098803c1171bbdbec5
Resolves: #28063
Reviewed-on: http://review.typo3.org/3254
Reviewed-by: Andreas Wolf
Tested-by: Andreas Wolf