[BUGFIX][Cache] File backend: Create directory after flush()
The file backend does not re-create the cache directory on
flush() which leads to problems if calling flush() and set()
afterwards again. One unit test had to manually clear cache
files because of that and can be simplified now.
This is an issue in TYPO3 v4 only, FLOW3 is not affected.

Change-Id: Ic7756b2c14641789696f10c223d76439ca7e9602
Resolves: #28090
Reviewed-on: http://review.typo3.org/3249
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn