[BUGFIX] Problem with PHP function ImageTTFBBox
A PHP bug lets ImageTTFBBox return wrong negative values in some cases.
That causes that sometimes images are not build correctly, but only with
a width of 1px. I know, it's a PHP bug and not a TYPO3 bug, but there's
an easy workaround and the problem is solved for all TYPO3 users with
affected PHP version.

There's an issue according to that problem on forge
(http://forge.typo3.org/issues/21054) but I didn't find a corresponding
report on Mantis.

Resolves: #21054
Releases: 4.6, 4.5, 4.4, 4.3

Change-Id: I6c1548e0f8762d6c2befdd47dbfb7810b01c9508