[BUGFIX] Fix front end redirect to 1-2-3 installer
An empty value in the TYPO3_db constant indicates a not configured installation in 
the front end. The front end then redirects to 1-2-3 installer. 
This broke with change 89732c4a of #27954 because the check for db constant 
is done too late.

Change the order of the checks during database connection and 
check if the TYPO3_db constant has a value first.

Change-Id: I0a7003d876a747e16552fecf6385b0449de5617a
Resolves: #28185
Related: #27954