[BUGFIX] Improve usability of group fields
If you set
    'disable_controls'=>'list'
you can _once_ upload an image (or file). But because of a missing Delete-Button
you never are able to change this image (or file) later.

Solution is, to add the delete Button, even if the list is not shown.
To remove even the delete button, use
    'disable_controls'=>'list,delete'

Change-Id: Ia13a9a25a64742eb49e64b7f26c085f15ca72b87
Resolves: #22679, #M14430
Releases: 4.6