[FEATURE] Speedup Backend with Cardlayout
This patch adds a cardlayout to the Backend.
BE Modules may register to the pagerenderer and add cards directly
to the layout.

This way changing the module is quite fast.
The extensionmanager or the about module can be simply reworked to support
that featuere.

Other new features like the dashboard depend on this change too.

Change-Id: I2bf206b7154c20575b0d2ce72150621b2dfbdb35
Resolves: #12664
Releases: 4.6