[BUGFIX] Unit test checkDateWithInvalidDateValues fails with PHP warning
If TYPO3 is configured to not suppress PHP warnings, the unit test
'checkDateWithInvalidDateValues with data set "Not Good"'
of Testsuite tx_scheduler_ModuleTest fails.

As it tries to insert non-numeric values into mkstring(), this warning
is raised.

We can safely suppress the output of PHP warnings for the mktime call,
as afterwards its return value is checked.

Change-Id: I7c6b9f587b574334e0096c800de98cb6ffc2033f
Resolves: #28271