[FEATURE] Add hook to tslib_fe->getHash()
This hook enables you to modify the hash_base that is used to store
different versions of a page in the cache. Normally it is fixed to
take the id, cHash, groups and some other stuff into account, but does
not allow to add own parameters.
Using this hook you can enable to store different versions of a page
depending on other parameters than GET-parameters.
It also solves an issue that different values of linkVars will produce
the same page if used independent from TypoScript conditions and without
cHash (which otherwise should be forced for all plugins and menu
generation).
The modified patch cleans up the hashbase creation and allows
to use the hook for the lockhash as well.

Change-Id: I68b0cfbcbec7eee496c4f903e3a129b7f33ab7c4
Resolves: #28299
Releases: 4.6