[FEATURE] Enable the GIFBUILDER cObject to render 32-bit PNG images
Introduce the "backColor" value "transparent" to create a PNG file
that has a transparent color as background and enhance the existing
code to keep it until the final image is written to a file.

If "transparent" is set the following conditions must be met:
* png_truecolor must be set to TRUE
* reduceColors must not be set
* format must not be set to jpg or gif

If one of those conditions is not met, GIFBUILDER falls back to
its default behaviour. Already existing backColor features are not
affected by this change.

The options "transparentBackground" and "transparentColor" have no
effect if the backColor is transparent

Change-Id: I0ce9a891710bbe1e7cae935c765aff2aa5640415
Resolves: #28405
Releases: 4.6