[FEATURE] Backport t3lib_utility_Version
Deprecating t3lib_div::int_from_ver() in 4.6 leads to major developer
annoyance because this method is used to switch to compatibility code
for older versions of TYPO3 in extensions and a hocus-pocus is needed
to prevent calling it if using 4.6.

Prepare this major deprecation by backporting
t3lib_utility_Version::convertVersionNumberToInteger() but without
refactoring existing code to use it.

Change-Id: I72673a483e5e716b24ebb6a89aaee3e09aaa7e65
Resolves: #28499
Releases: 4.5