[TASK] Prepare deprecation of t3lib_div::int_from_ver()
Deprecating t3lib_div::int_from_ver() in 4.6 leads to major developer
annoyance because this method is used to switch to compatibility code
for older versions of TYPO3 in extensions and a hocus-pocus is needed
to prevent calling it if using 4.6.

Deprecate this method in two phases:
- TYPO3 4.6: mark it as deprecated
- TYPO3 4.7: use deprecation log to deprecate it really

This method should only be removed in TYPO3 4.9.

Change-Id: I73cfc9a5e15f6f336534e3705e914fc37862472e
Resolves: #28499
Releases: 4.6
Reviewed-on: http://review.typo3.org/3733
Reviewed-by: Benjamin Mack
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers