[FEATURE] htmlArea RTE: Add possiblity to set pastetoggle to ON by default
Two new Page TSConfig properties are added:

RTE.default.buttons.pastetoggle.setActiveOnRteOpen : If set, and if the pastetoggle button is enabled, the button is toggled to ON when the RTE opens.

RTE.default.buttons.pastetoggle.hidden : If set, and if the pastetoggle button is enabled, the button is hidden in both the toolbar and the context menu. Hence, if buttons.pastetoggle.setActiveOnRteOpen is also set, all paste operations will be performed using the current clean paste behaviour.

The manual is updated to document the fact that default and/or overriding values may be set for the clean paste behaviour. For example: setup.override.rteCleanPasteBehaviour = plainText

Change-Id: I3f479846d5ce373f7f386fd450603faed08f7e8d
Resolves: #25039
Releases: 4.6