[BUGFIX] JSwindow-typolink URLs are encoded twice
In #28189 a bug was introduced for JSwindow-typolink URLs.
Because t3lib_div::quoteJSvalue is called without second paramter,
the value passed to it will also be encoded with htmlspecialchars()
which is then done a second time when putting the link together.

Change the t3lib_div::quoteJSvalue call to not encode
the result with htmlspecailchars()

Resolves: #28536
Related: #28189

Releases: 4.3, 4.4, 4.5, 4.6

Change-Id: If6bcad4d2bf1e0a4b3f5a6f04e4718df8e4e7150