[BUGFIX] Autoloader Cache is not updated
The autoload registry contents are stored in a separated PHP
level caching in TYPO3 4.6. However, during development (and
also maybe during TYPO3 Source Updates) new classes in the
registry files are not considered.

This fix introduces a file contents check by creating a hash
for all valid autoload registry files that will be used with
the cacheIdentifier.

Change-Id: I4fa47178016925f39ff2168f65afb1ae357be556
Resolves: #28589