[BUGFIX][Cache] Let NullBackend implement PhpCapableBackend
The null backend is very useful to effectivly disable caching
during development. With this change it also implements the
PhpCapableBackend interface and can be used to replace the
FileBackend this way.

This is the v4 backport of FLOW3 changeset
https://review.typo3.org/#change,3910

Additional to the FLOW3 version the null backend is now
required in config_default.php to enable the cache_phpcode
to use this backend in autoloader for development context.

Change-Id: Icca11825742eaa64b1e6f2db2aec48590887cb0b
Resolves: #28598
Related: #28589
Reviewed-on: http://review.typo3.org/3911
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Susanne Moog
Tested-by: Susanne Moog