[BUGFIX] htmlArea RTE: Relative url's on pasted links/images in Firefox
When copying/pasting an internal link or image from and to the RTE
editing area, Firefox changes the href or src to a relative url.
After save, the url ends up being wrong.

Similar issues arises when moving images or links inside the editing
area and when dropping images in the RTE from the image dialogue
window.

This change works around Firefox issue
https://bugzilla.mozilla.org/show_bug.cgi?id=613517,
by making the url absolute again before saving or toggling
to source text mode, as well as on paste and drop events.

Change-Id: I3b43291b9c8fa50763d649ac0fdc9b9e692869bf
Resolves: #28497, #24852, #27353
Releases: 4.4, 4.5, 4.6