[FEATURE] Make filelinks icons configurable
The changeset introduces new Typoscript properties for setting the path
to the filetype icons, the extension of the filetype icons and
the width and height attribute of the img tag when using
the Filelinks/"uploads" cObject.

If no special configuration is set, the old defaults are taken (icons
from tslib/media/fileicons/, extension gif, width 18, height 16).

New sub properties of filelink.icon:
* path : path to icon set (default: tslib/media/fileicons/)
* ext : extension of icons (default: gif)
* widthAttribute : width of icons (default: 18)
* heightAttribute : height of icons (default: 16)

Change-Id: I69781a5ece9ff32ae51220e669b0c740638f5329
Resolves: #23624
Releases: 4.7
Reviewed-on: http://review.typo3.org/3921
Reviewed-by: Markus Klein
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich
Reviewed-by: Simon Schaufelberger
Tested-by: Simon Schaufelberger
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter