[FEATURE] Make filelink icons configurable
The changeset introduces new Typoscript properties for setting the path
to the filetype icons, the extension of the filetype icons and the
width and height attribute of the img tag when using the
Filelinks/"uploads" cObject.

If no special configuration is set, the old defaults are taken (icons
from tslib/media/fileicons/, extension gif, width 18, height 16).
TSref needs to be updated to reflect the new properties.

See Forge for TS properties.

Change-Id: I69781a5ece9ff32ae51220e669b0c740638f5329
Resolves: #23624
Releases: 4.6