[FEATURE] Indexed Search: Add additional extbase plugin
Introduces a new plugin (pi2) with fluid and extbase that
separates the DB logic (Repository) and the output
(fluid-based now) from the rest (searchcontroller).
However, there is no real "repository" that works as
a regular repository. Additionally, some options
have been renamed.
This is still not completely tested and a work-in-progress,
but the feature-set should be complete. This feature is also
ready for the MySQL Fulltext Indexed Search.

Change-Id: I8ed4a3e5f09db64d560d0a44b4a7c51c6ec80481
Resolves: #28612
Releases: 4.7
Reviewed-on: http://review.typo3.org/3926
Reviewed-by: Stefan Neufeind
Reviewed-by: Dominik Mathern
Tested-by: Dominik Mathern
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter