[FEATURE] Show warning message if the open record changed during editing
This bugfix adds a polling algorithm to the tceforms to check,
wether a record is changed, it lowers the
increased risk of a lost update with the new cardlayout.

The change is detected by using the lastmod timestamp of the related
record.

Logic of this change:
1. onReady
2. find formElements named according tceforms convention and create
   list of edited records
3. initial poll for lastchange
4. periodical polling

After each polling the lastchange time is checked for changes, having a
change you get an Ext.Msg.confirm to decide, wether to reload the form
or merge the changes manually.

Change-Id: I89119ea6bed827b1cb44f0c6380a13961f328697
Resolves: #28618
Releases: 4.6, 4.5