[BUGFIX] Fix HTTP Request Exceptions with SSL certificates
HTTP_Request2 expects the default value for ssl_cafile, ssl_capath,
ssl_local_cert and ssl_passphrase to be NULL, and not ''.

As the Install Tool is not able to handle this, we need to check if
those values are empty and set them NULL accordingly.

Additionally the socket adapter has trouble with ssl_verify_peer.
Therefore change ssl_verify_peer to FALSE by default

Change-Id: I065c649a671d4360c9a51ab4be40fb0eb58d6ec6
Resolves: #28619
Reviewed-on: http://review.typo3.org/3949
Reviewed-by: Stefan Neufeind
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers