[BUGFIX] t3lib_tcemain unset empty fields even if a lower range is set
If a lower limit was set in the TCA configuration of a field (for example,
today's date for an endtime field of tables created with kickstarter),
this field could not be cleared once a value has been entered.
The empty value was replaced with the lower limit as soon as the record
was saved.

TCE now checks if the value is an empty string before comparing with
the upper and lower ranges.

Change-Id: If9b26ef59a061cde006ba97bd9ea4ef5dc5b86b4
Fixes: #26503
Releases: 4.5, 4.6