[TASK] Disable compression for inlineJs in Backend
JSMin is a major performance nightmare and no network connection can be
as slow as it would justify to spend so much time to execute JSmin
several times for each and every backend request.

As long as we have no fast replacement in core available, disable
compression for all JS snippets added with
t3lib_PageRenderer::addJsInlineCode in the Backend.

JSMin's share of execution time was about 20-25% of the page generation
time, for almost empty backend pages even ~50%.

Change-Id: Iabbd092a349dec6e84c40b7b7c750bc718168d85
Resolves: #29498