[BUGFIX] Maintain compatibility with changed headline rendering
If the fontTag property is set and the dataWrap property is set to the
default value, replace the dataWrap with the fontTag property value and
disable insertData on this level (if set).

This is to retain compatibility with versions before 4.5.4 while
compatibility with modified templates (before and after 4.5.4) is still
provided.

Change-Id: Ieffeed7b7d766b0d248ed666bfef6e8f62ea1f38
Resolves: #28847
Related: #26876
Releases: 4.5, 4.4, 4.3
Reviewed-on: http://review.typo3.org/4283
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader
Reviewed-by: Michael Stucki
Tested-by: Michael Stucki