[BUGFIX] Handle cache tables in install tool and em
This patch moves the needed table definition for a cache from the caching
framework to the install tool and the extension manager. This way the cf
tables are handled just like every other table and show up for example in
compare wizard of install tool and if an extension is installed that
registers own caches (like extbase).

Change-Id: I61e7dd1b1fbe970529810876bc7137986f9ae656
Resolves: #28973