[BUGFIX] Enlarge on click for images does not open a popup
With issue #27983 some deprecated parts were removed but obviously too much. Although $GLOBALS['TSFE']->additionalJavascript is used throughout the core, after this issue it was never rendered anymore in the frontend.

The javascript function openPic was never added to the frontend because of this.

This patch reinstates the responsible lines which were removed for Enlarge on click

Change-Id: I8e182be8e894e49212a5a4249118b44c0e0bd954
Resolves: #29041
Releases: 4.6