[BUGFIX] TYPO3 tries to connect to typo3.org on every login
If the extension cache is disabled,
the extension manager tries to establish a connection to the TER
as soon as user credentials for a user are given.

This patch changes that behavior so that it is only done for
the functions that actually need a TER call.

Additionally I did some cleanup.

Change-Id: I1633334d762d713fcf39097192ed6fa8fcfcf343
Resolves: #28949
Releases: 4.6, 4.5