[BUGFIX] Remove check for invalid prefix in addService
t3lib_extMgm::addService() checks that $serviceType (eg. 'auth') does not start
with a valid class prefix. While it is probably not a good idea to prefix a
service type with 'user_' or 'tx_', there is no technical reason to enforce
this. The check especially leads to problems if
TYPO3_CONF_VARS['FE']['userFuncClassPrefix'] is set to an empty string.
This patch removes the check.

Change-Id: I078078eef5147cb267483a0d8d04ab113a24f7e9
Resolves: #29165
Releases: 4.6, 4.5