[BUGFIX] Remove check for invalid prefix in addService
t3lib_extMgm::addService() checks that $serviceType (eg. 'auth') does not start
with a valid class prefix. While it is probably not a good idea to prefix a
service type with 'user_' or 'tx_', there is no technical reason to enforce
this. The check especially leads to problems if
TYPO3_CONF_VARS['FE']['userFuncClassPrefix'] is set to an empty string.
This patch removes the check.

Change-Id: Iaf60da0eae61a2c64bb4626f5629b8c262d00cc2
Resolves: #29165
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/4528
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn