[TASK] Streamline class prefix handling in the core
At the moment the frontend option userFuncClassPrefix is also checked in
various backend scenarios (for example the registry) and in general
everywhere t3lib_div::hasValidPrefix is called. Therefore it should be a system
option. Additionally to ensure "user_" is working as a prefix this was given
hardcoded as additionalPrefix to the function in some cases.

This patch streamlines the handling of prefixes, introduces a [SYS] option
additionalAllowedClassPrefixes, deprecates the [FE] option and allows
the prefixes 'tx_', 'Tx_', 'user_' and 'User_' by default.

Change-Id: I70fa89120dbe8c7790e3d6de4b173b69beb8a6f9
Resolves: #29166
Releases: 4.6
Reviewed-on: http://review.typo3.org/4526
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn