[BUGFIX] 'Update settings' button does not work in new em
After installing an extension you get a popup where you can
update the configuration settings of an extension. In the new
extension manager the update button there will not work as it
is a submit button but the form around the fields is missing.

Attached patch simply inserts the form in this case, too.

Change-Id: I9c4d5c1f1b84f1b6d7c3e1e1be33ee63752d09e9
Resolves: #29172
Releases: 4.5, 4.6