[BUGFIX] 'Update settings' button does not work in new em
After installing an extension you get a popup where you can
update the configuration settings of an extension. In the new
extension manager the update button there will not work as it
is a submit button but the form around the fields is missing.

Attached patch simply inserts the form in this case, too.

Change-Id: Iaa2eb8876e96bbe7d3120cd124c37d29815647fb
Resolves: #29172
Releases: 4.5, 4.6
Reviewed-on: http://review.typo3.org/4629
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn