[BUGFIX] Fix misleading SQL error message in TCEforms
When an SQL error occurs, TCEforms shows e error message "This may
indicate a table defined in tables.php is not existing in the database!"
The error can be caused by any schema mismatch, not only missing tables.

The error message is now localizable and points to the database compare.

Change-Id: I31bea64cf656c951bc63a89eff47fbeb893ae73f
Resolves: #16306
Releases: 4.6,4.5