[BUGFIX] Unlock Install Tool only after acknowledgement
Since #28621, the Install Tool gets automatically unlocked, if an admin
user opens it from the Backend.

This reduces the admin's awareness of the Install Tool being
explicitly protected by the ENABLE_INSTALL_TOOL file. This change
introduces an additional step, in which the user is explicitly asked for
confirmation to unlock the Install Tool.

Change-Id: I289959f7cdc8ff05221005d4431f2e27b3f5cf95
Fixes: #28689
Related: #28621
Releases: 4.6
Reviewed-on: http://review.typo3.org/4777
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Jigal van Hemert
Tested-by: Jigal van Hemert