[FEATURE] Allow user-defined additional backend languages
At the moment, any language may be used for Frontend output. But in Backend, the list
of supported languages is hardcoded at various places. This prevents additional
languages to be defined locally, for instance when translating TYPO3 into a new
language, before it gets officially supported by the Core.

Change-Id: Ibd3236a82790f9f20c91ff3663d6625694ef7cf6
Resolves: #29513
Releases: 4.6