[BUGFIX] Sanitize identifier when compiling templates
The "identifier" of a compiled template is based on the
last modification timestamp and on the name of the original
template/partial/layout.
if that name contains special characters (like dashes),
the class name of the compiled template is invalid
resulting in a fatal error.
This is fixed by sanitizing the identifier before creating
the compiled template.

Change-Id: I99947683923175d890077a5f98ad8922e24e60b3
Related: #29457