[FEATURE] Allow user-defined additional backend languages
At the moment, any language may be used for Frontend output. But in Backend, the list
of supported languages is hardcoded at various places. This prevents additional
languages to be defined locally, for instance when translating TYPO3 into a new
language, before it gets officially supported by the Core.

Change-Id: I66cd9149c860746356e249c778a5303211076d37
Resolves: #29513
Releases: 4.6