[TASK] EM: maintenance tab suggest download of CF tables
When looking at an extension's details, the developer information tab
lists all caching framework (CF) tables and the maintenance tab
proposes all CF tables for download. It doesn't make sense, even
for cache tables that are related to the current extension
(because they are created automatically and there's no reason
to dump a cache table's data).
The solution is to add a new flag to avoid listing CF tables
in some circumstances.

Change-Id: I3a8d43bc0d94d9bca99ae72a808e8e07057e9852
Resolves: #29592
Releases: 4.6
Reviewed-on: http://review.typo3.org/4886
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers