[FEATURE] EM: Make full-screen mode configurable
Some windows could be maximized for a better overview.
Since this may be troubling for some users, introduce
a configuration parameter for explicitly enabling
full-screen mode.
Also allow the user to maximize or minimize the
modal window, independent of the setting.
Unify width and height, where it makes sense.

Change-Id: Ibe5a76bccdeddc9c250cdcecaaf12a0c26105847
Resolves: #29630
Releases: 4.7