[BUGFIX] Don't add uid,pid if  <table>.'*' is present in field list.
tx_cal_tsservice uses queries with select <tablename>.* and complex joins.
This breaks due to  the adding of uid,pid if the query is set.

Do not add uid,pid,... to field list if the field list already contains a
'*' or the repective field is in the list.

The regexp matching is necessary to  detect fields like post_uid
(comments extension).

Change-Id: I50332c22e627ea452aaee233fdbbdf3dd426a1b6
Resolves: #29783
Releases: 4.6