[BUGFIX] t3lib_Compressor: Avoid reference in foreach loop
This fixes a bug introduced in #29695. There is a strange php bug
when using $filename as a reference and later on as a variable in
a foreach loop. The problem is described here:

http://www.php.net/manual/de/control-structures.foreach.php#101425

PHP 5.3.2 (used in current Ubuntu) and 5.3.3 (used in current Debian)
are affected. Other versions could be affected, too.

Avoiding the use of reference in the for loop fixes the bug.

Resolves: #29832
Releases: 4.6
Related: #29695

Change-Id: I74c0ceee40e170b7cf2e7db91b1554d811a7c844