[TASK] Index for "pages" table is still not optimal
The fix for #21381 causes "using filesort" in the List view for
"pages" table. This happens due to the "hidden" field in the
index. This field should be removed from the index because it is
not needed for the List module and it is not useful in FE due to
more expensive fe_group field.

Change-Id: Ia5fc87f7b1fa89e515c292248365d1d684872be0
Resolves: #29896
Related: #21381
Releases: 4.6