[BUGFIX][SECURITY] Arbitrary TypoScript execution on system extension form
The new system extension form can be used to render custom FORM elements as
well as regular cObjects like TEXT or COA. Since the form wizard can be used
by any editor in the backend and writes data to the field bodytext, this can
also be used to execute arbitrary TypoScript without further access checks.

This change introduces two defined and allowed content elements "header" and
"textblock" that can be defined by using the form wizard. If the TypoScript
that was generated by the mentioned wizard is rendered, regular cObjects are
disabled. If the FORM or FORM_INT cObject is used directly from a TypoScript
template, all possible cObjects can still be used.

Change-Id: I573764de7583b078456e71e95ea7903b433c29db
Resolves: #30095
Releases: 4.6
Reviewed-on: http://review.typo3.org/5128
Reviewed-by: Andreas Wolf
Reviewed-by: Frederic Gaus
Tested-by: Frederic Gaus
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader