[BUGFIX] Rewriting localconf.php endlessly expands lines
When the file localconf.php gets rewritten, and an extension's
configuration value contains two subsequent slashes ("//"), the part
from after the slashes until rest of the line is placed at the end of
the newly generated comment. This is repeated each time localconf.php
is written, so the comment is extended again and again. The rewriting
does not take into account that "//" can be part of a value, but thinks
that everything after the first "//" is.

Change-Id: I58d392025d88af58ea1c069fef050251da5859ee
Resolves: #29252
Releases: 6.0, 4.7, 4.6, 4.5