[BUGFIX] *IFSUB ignores l18n_cfg field
The l18n_cfg field is used to prevent the appearance of the default
language record or any alternative language records that haven't a translation
in the menu. At the first glance it seems to work like expected, because the
expected page records are not visible in a configured frontend menu. If we have a
closer look at this part, we will find out that the flag is badly ignored for the
record states *IFSUB. This causes that a page record which has no visible
subpages is marked with the hasSubmenu flag.

Change-Id: I9a6be9ca3c880574d088cbc7bf8795cc9b2de6d3
Resolved: #19925
Releases: 4.6, 4.5, 4.4