[BUGFIX] *IFSUB ignores l18n_cfg field
The l18n_cfg field is used to prevent the appearance of the default
language record or any alternative language records that haven't a translation
in the menu. At the first glance it seems to work like expected, because the
expected page records are not visible in a configured frontend menu. If we have a
closer look at this part, we will see that the flag is badly ignored for the
"*IFSUB" record states. This causes that a page record which has no visible
subpages is marked with the hasSubmenu flag.

Change-Id: Ia988dfeb272b6c64c86c45f9d80b8b9299a09bfc
Resolves: #19925
Releases: 4.7, 4.6, 4.5, 4.4
Reviewed-on: http://review.typo3.org/5573
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Stefan Galinski
Tested-by: Stefan Galinski