[BUGFIX] try-catch avoids exception side effect in "template setup".
At least in IE9, "if (!this.mirror.editor)" throws an exception in case of
"editor" is undefined (depending on browser settings, this may be silent).
With "client browser caching" set to an option other than "read always",
a first touch on "template setup" is rendered as expected, but any further
touch on "template setup" (within the same session) leads to a incorrect
display, due to the thrown exception.

Change-Id: I238c2e7297f162c91dacbece044589cc941b3de1
Resolves: #30764
Release: 4.3