[BUGFIX] Use select query to check for existing pagesection cache
Checking for existing pagesection cache is done by performing an update
query and checking if the number of resulting rows is > 0. Not all databases support this.
When using Oracle with OCI8, doing an update on a non existing record will throw a warning.
This is solved by doing a select first.

Change-Id: I7e5cf78a30d5a88c49c52acf15188872053a6ba1
Resolves: #30544
Releases: 4.5