[BUGFIX] Reports module: Available services broken
Displaying the available services in the reports module is broken.
This is a regression of removing the global declarations for
TYPO4 4.6. In TYPO3 < 4.6 the global array $T3_SERVICES contained
too much information:

$T3_SERVICES[$serviceType][$serviceKey]
   --> Information about a particular service
$T3_SERVICES[$serviceKey][$serviceKey]
   --> Same information

Because of this duplicate, the method t3lib_extMgm::findService could
be "abused" to find out if a particular service was the service of
highest priority of a type.

After the cleanup, the array $GLOBALS['T3_SERVICES'] only contains
the information it is supposed to:

$T3_SERVICES[$serviceType][$serviceKey] --> Information about a
particular service

Therefore t3lib_extMgm::findService($serviceKey, '*') will no longer
work. This can be easily solved by adding a method that checks, if a
particular service is the service of highest priority for the given
service type.

Change-Id: I15fd7eecaafb52df0599598c5b700e458e63f0d9
Resolves: #30554
Releases: 4.6