Revert "[FEATURE] Indexed Search: add mySQL fulltext support"
This reverts commit 6294f1fa8fd13a27ce93cfe6257d6754e8e0e1a6

Although I gave my +1 by reading a while ago, this did not mean this patch was ready to be merged and for sure not with a single review by testing. I gave this review because I started to give my opinion on this patch and I had a few comments to be tackled with. I did not pay (enough) attention to the 4.6 target of the commit message.

I really appreciate the work that has been done by all of you but I cannot accept such a big patch fly into the Core so late according to the roadmap (RC in just a few days) and above all to the feature freeze date that was prominently advertised since the very beginning.

Sorry but this is a clear no-go to include this new feature at this stage. It will have to be properly tested and integrated within TYPO3 4.7.