[BUGFIX] TCEforms draws huge empty icon row which on select-fields
Certain fields of type "select" which offer records from a
"foreign_table" will contain a huge amount of t3-icon-empty
icons just below the select box. If you have enough of them,
you even get a horizotanl scrollbar.

See some screenshots here:
http://forge.typo3.org/issues/13422

The issue is not a CSS issue, but a bug in the rendering
of such a field.

See it in action in tt_content field "sys_language". Just create
enough sys_language records so that you see the row of empty
icons being created. This wasn't this way before.

Change-Id: I2e2e59c906b2e239ba161b16aa1980f454210229
Resolves: #25398
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/7118
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer