[BUGFIX] TCEforms draws huge empty icon row which on select-fields
Certain fields of type "select" which offer records from a
"foreign_table" will contain a huge amount of t3-icon-empty
icons just below the select box. If you have enough of them,
you even get a horizotanl scrollbar.

See some screenshots here:
http://forge.typo3.org/issues/13422

The issue is not a CSS issue, but a bug in the rendering
of such a field.

See it in action in tt_content field "sys_language". Just create
enough sys_language records so that you see the row of empty
icons being created. This wasn't this way before.

Change-Id: I664400ee6ee1d778e227b1858f79b9ed98482d0a
Resolves: #25398
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/5569
Reviewed-by: Stefan Galinski
Tested-by: Stefan Galinski
Reviewed-by: Dmitry Dulepov
Tested-by: Dmitry Dulepov