[BUGFIX] Sporadic memcache error on server load
Memcache server sometimes closes connection under heavy load. In
such cases TYPO3 shows error messages in Frontend when setting
the value to cache fails. This should not happen because it is
not such a error that should cause server down times. Change the
exception to a warning in sys_log instead. Admins can see the
error and work on it.

Change-Id: I43e9356658d06f9101e1b488519ec76af6319d2d
Fixes: #30162
Releases: 4.7, 4.6, 4.5
Reviewed-on: http://review.typo3.org/7411
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers