[BUGFIX] Sporadical memcache error on server load
Memcache server sometimes closes connection under heavy load. In
such cases TYPO3 shows error messages in Frontend when setting
the value to cache fails. This should not happen because it is
not such a error that should cause server down times. Change the
exception to a warning in sys_log instead. Admins can see the
error and work on it.

Change-Id: I3d461ef2517d8d10a161c1a1404931543681bb97
Resolves: #30162
Releases: 4.7, 4.6, 4.5