[TASK] Remove unused option for "oncontextmenu"
The TYPO3_CONF_VARS option to enable oncontextmenu is not
needed anymore IMHO, as this option is official part of
HTML5 (previously it wasn't allowed by IE for HTML4, and a
proprietary option for Firefox (back when we introduced it
in 4.1). The patch removes any occurence of this option and
enables oncontextmneu everywhere.

Change-Id: I487accb5717af79072c704fde34018827850269d
Resolves: #30613
Releases: 4.6