[TASK][DB] Followup: Move workspace preview into versioning
Followup to: Iee2f4ea42aa28825d991df92eff53c7143d93e83
as there was one changed patch file missing on the last
patchset.

As the whole handling of the ADMCMD_preview is related to
version / workspaces, the according code should go in
these extensions, thus it is removed in the core.

Change-Id: I8931dbf642a98c259715a16a126aad3d23a9785d
Resolves: #27931
Releases: 4.6