[BUGFIX] handler is undefined when resizing the backend
Resizing the backend browser window results in a JavaScript error.
The function that handles the re-positioning of the menus does not
work correctly. The observer requires a reference to a callback
function - currently the method was directly executed and the return
value used as function reference, which was undefined in this case.

Change-Id: I15f3bab4c3c59c5c96633423ac7cd0ebd1736a1a
Resolves: #25787
Releases: 4.6, 4.5