[BUGFIX] Race-condition while updating extension list
There is a race-condition in the EM while updating the extension list.

If the update takes more than 30 seconds, Ext.Direct reaches its Ajax
timeout and sends the Ajax request (to update the list) again. As the EM
updates the extensions.xml.gz file at the beginning of the update
process, the second request immediately sees that the repository is
already up-to-date and respons this status to the client.

Thus, the user gets a "Repository is already up-to-date" message shown,
although it was not.

Change-Id: I2bc6b83f0951e16b895607025c2d3bad392b9759
Resolves: #30673
Releases: 4.6, 4.5