[TASK] Move FEworkspace code into versioning / part 2
As there is a hook after initializing the BE user,
the code that initializes the BE user when using
a previewlink can now be moved out of the core.

Also the same hook can now be used to do the generic
workspace initialization.

Additionally, the remaining functions in TSFE
should be cleaned up.

Change-Id: I1674c89bc195f2f1b154574fefa57c16bd8b2385
Resolves: #30682
Releases: 4.7
Reviewed-on: http://review.typo3.org/5644
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch
Reviewed-by: Stefan Neufeind
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Wouter Wolters
Tested-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter