[BUGFIX] Showmodule cannot be called before Store is loaded
Example: Using the direct page edit feature (backend.php?edit=xxx)
the BE responds with "Module loader No module found. If this is a
temporary error, please reload the Backend!" error.

Since the AJAX request is not finished at the time this is called,
it has to be checked whether the store was completely loaded.

Change-Id: I555427d61845678c1f90e2039c41cdd2055e5b6a
Releases: 4.6,4.5
Resolves: #28951
Reviewed-on: http://review.typo3.org/5712
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers