[BUGFIX] Showmodule cannot be called before Store is loaded
Example: Using the direct page edit feature (backend.php?edit=xxx)
the BE responds with "Module loader No module found. If this is a
temporary error, please reload the Backend!" error.

Since the AJAX request is not finished at the time this is called,
it has to be checked whether the store was completely loaded.

Change-Id: Ia7c39fc94e75ef003a7a839856ea78cbd0854a66
Releases: 4.6,4.5
Resolves: #28951
Reviewed-on: http://review.typo3.org/5681
Reviewed-by: Helmut Hummel
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers
Tested-by: Jigal van Hemert