[BUGFIX] handler is undefined when resizing the backend
Resizing the backend browser window results in a JavaScript error.
The function that handles the re-positioning of the menus does not
work correctly. The observer requires a reference to a callback
function - currently the method was directly executed and the return
value used as function reference, which was undefined in this case.

Change-Id: I9006793a2450b7b8fee408e8ef5668d359138390
Resolves: #30760
Related: #25787
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/5695
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader