[BUGFIX] Fix for »Fatal error: Call to a member function offsetExists() on a non-object in xyz« in pagetree topPanel filter
Setup:
------
Typo3: 4.5.6 (fresh installation)
PHP: 5.2.13 (cgi) and PHP 5.3.5 (cgi-fcgi)
mysql: 5.1.39
Host: Domainfactory managed
Three domains/sites nearly 1300 pages

the exact reason is unknown to me but if you go to the pagetree topPanel
filter as shown in the file and type a certain keywort (in this case
disount; but also has also reported with other keywords) you get the
following error message:

Fatal error: Call to a member function offsetExists() on a non-object in
typo3_src-4.5.6/t3lib/tree/pagetree/class.t3lib_tree_pagetree_dataprovider.php on line 180

Also typo3 backend behaves really strange after this happens.

The attached patch fixes this behaviour. I am not sure though if it is
just a workaround or if it ist is a valid fix. Maybe it is a
missbehaviour in t3lib_tree_pagetree_NodeCollection?! Anyway an
assertation that $reference really is an object before you call
offsetExists seems legit to me.
Change-Id: I3734769a068e442359dc69d4262fd237f9384860
Fixes:
Releases: 4.6