[TASK] Add CSS Class to special input types in FORM
All input elements are wrapped by a list element that only has one
css class containing the element id of the field. To avoid unneccesary
barriers to ie. hide inputs field of the type "hidden" an extra class
is added to the list element to determine the field type of the
wrapped element, ie. hidden, text, ... .

Change-Id: I4f39e4663f921591b55d1f020f314fe7e076e0ad
Resolves: #30851
Releases: 4.6, 4.7