[TASK] Add CSS Class to special input types in FORM
All input elements are wrapped by a list element that only has one
css class containing the element id of the field. To avoid unneccesary
barriers to ie. hide inputs field of the type "hidden" an extra class
is added to the list element to determine the field type of the
wrapped element, ie. hidden, text, ... .

Change-Id: Ia14641d488f10e2a60f9ba0fafdc10f31d0eeb1d
Resolves: #30851
Releases: 4.6
Reviewed-on: http://review.typo3.org/5745
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Stefan Neufeind
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter