[BUGFIX] Class not found fatal in autoload
If the caching framework configuration is broken and cache_phpcode
array is not set properly, cacheManager throws a fatal error unable
to find class t3lib_cache_exception_NoSuchCache. The patch adds the
class to be required during bootstrap so the exception (uncaught) is
actually thrown and hints an administrator on what exactly goes wrong.

Change-Id: I42ac66457776e1b4a4f8dca40e41f364d9496eed
Resolves: #30913