[BUGFIX] Sorting of localized records is wrong after copying
Solution: When copying a record in original language, if a localized
record of the destination record exists for language A, then the
localized record of the copied record, if it exists for language A,
should be inserted after the localized record, in language A, of the
destination record.
The same applies when moving a record in original language.
Similar situation arises when a new localization record is created.

Change-Id: If01e602b0fe68423baba9f9973bd5237c1d0b75a
Resolves: #30469
Releases: 4.5, 4.6, 4.7
Reviewed-on: http://review.typo3.org/6032
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Stefan Neufeind
Reviewed-by: Tolleiv Nietsch
Tested-by: Tolleiv Nietsch