[BUGFIX] Sorting of localized records is wrong after copying
Solution: When copying a record in original language, if a localized
record of the destination record exists for language A, then the
localized record of the copied record, if it exists for language A,
should be inserted after the localized record, in language A, of the
destination record.
The same applies when moving a record in original language.
Similar situation arises when a new localization record is created.

Change-Id: I7e95268e159f5c652c1c99b9663293e3a7ff9096
Resolves: #30469
Releases: 4.5, 4.6, 4.7