[BUGFIX] Menu link creation should always return an array
According to the documentation, method link() returns an array with a
few keys. In case of error however, FALSE is returned which in turn
crashes PHP if the returned value is not post-processed and sent as
this to t3lib_div::implodeAttributes() which forces an array to be
passed as argument.

Change-Id: I7a593dd7eec2bee9e9f84415ccdd592f4d767caf
Fixes: #31079
Related: #17037
Releases: 4.6