[BUGFIX] Check for rsaauth being usable
The Extension Manager executes shows the results of some checks in the
Configuration section of the extension "saltedpasswords", whether
"rsaauth" is loaded and whether it is enabled.

However, it lacks a check for "rsaauth" being really able to work. In
this case, the user must not be instructed to set his
"loginSecurityLevel" to "rsa", as login will be impossible afterwards.

Instead tell the user that OpenSSL extension is not available or not
working correctly.

Change-Id: Iacedf86dc75090120194f96b8f8ff00cdd8bc90e
Resolves: #31178
Releases: 4.6, 4.5, 4.4