[BUGFIX] Check for rsaauth being usable
The Extension Manager shows the results of some checks in the
Configuration section of the extension "saltedpasswords", whether
"rsaauth" is loaded and whether it is enabled.

However, it lacks a check for "rsaauth" being really able to work.
If "rsaauth" fails, tell the user that OpenSSL extension is not available
or not working correctly.

Change-Id: I80186c989d399fc88b601475a83637f0b71cbb3c
Resolves: #31178
Releases: 4.6, 4.5, 4.4
Reviewed-on: http://review.typo3.org/6267
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers