[BUGFIX] Check for rsaauth being usable
The Extension Manager shows the results of some checks in the
Configuration section of the extension "saltedpasswords", whether
"rsaauth" is loaded and whether it is enabled.

However, it lacks a check for "rsaauth" being really able to work.
If "rsaauth" fails, tell the user that OpenSSL extension is not available
or not working correctly.

Change-Id: Ie68b0f7dca4ceef9752cec44b1ff651e77f5f1d8
Resolves: #31178
Releases: 4.6, 4.5, 4.4
Reviewed-on: http://review.typo3.org/6268
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers