[BUGFIX] Check for rsaauth being usable
The Extension Manager shows the results of some checks in the
Configuration section of the extension "saltedpasswords", whether
"rsaauth" is loaded and whether it is enabled.

However, it lacks a check for "rsaauth" being really able to work.
If "rsaauth" fails, tell the user that OpenSSL extension is not available
or not working correctly.

Resolves: #31178
Releases: 4.6, 4.5, 4.4

Change-Id: Iacedf86dc75090120194f96b8f8ff00cdd8bc90e
Reviewed-on: http://review.typo3.org/6184
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Stefan Galinski
Tested-by: Stefan Galinski