[BUGFIX] EM: Popup window not scrollable
There are four places in the Extension Manager, when a JavaScript popup
is opened (not an ExtJS window).

These popups are created with
  window.open(..., width=650,height=500),
thus they have no scrollbars. While it is possible in most browsers to
scroll with a mouse wheel, this seems not to be possible with IE.

Change-Id: I71ed8393c24bc2a6e543b1e198ceb1bf240f1522
Resolves: #31184
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/6185
Reviewed-by: Steffen Gebert
Tested-by: Steffen Gebert
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers