[BUGFIX] Awkward styling of split bar in Chrome browser
Due to a recente change (#27660), the Ext JS split bars, which act as
separators between module menu / navigation and navigation / content
container, are increased in width only for the Chrome browser.

This breaks the styling and makes it look lovelessly implemented, when
the selected page from the tree should have a visual connection to the
content frame.

Change-Id: Iacca368168d122e9d2df05a22590d21859f6b3cb
Resolves: #31186