[BUGFIX] Fix misleading SQL error message in TCEforms
When an SQL error occurs, TCEforms shows the error message
"This may indicate a table defined in tables.php is not
existing in the database!"

The error can be caused by any schema mismatch, not only
missing table, so this is misleading.

The error message is now localizable and gives a hint to
run the database compare in the install tool.

Change-Id: I8083463f70dd27a83bb681577ba751b1b6044f2a
Resolves: #16306
Releases: 4.6, 4.5
Reviewed-on: http://review.typo3.org/6257
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers